// apps/web/src/lib/api/generated/models/registerDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Auth Tutorial API
 * API documentation for the auth-tutorial project. This is a comprehensive REST API built with NestJS.
 * OpenAPI spec version: 1.0
 */

export interface RegisterDto {
  email: string;
  name: string;
  /** @minLength 6 */
  password: string;
}